feat(admin): show variant preprocess tasks in /admin/tasks (reusable TaskRunPanel)
Extract a parameterized TaskRunPanel from the metadata-only tasks page and
render two of them under shadcn Tabs (Metadata | Variant preprocessing).

- TaskRunPanel owns the shared SWR wiring (runs / preview-count / selected
  detail), the kick-poll effect, the start/cancel handlers and the entire
  render (active run, history list, detail dialog). Per-task differences flow
  through a generic `config` prop: basePath, taskKey, defaultScope, a pluggable
  ScopeControl render-prop, previewCountQuery, scopeLabel, canStartWithoutPreview
  and mapStartError.
- Metadata tab is byte-identical to before: same album/show Select scope,
  same /api/v1/tasks endpoints, same polling.
- Preprocess tab uses a `{ force }` toggle, hits /api/v1/preprocess-tasks, and
  maps the 400 "Variant storage backend is not configured" error to a clear
  hint instead of a raw toast.
- Adds Tasks.* preprocess keys in all four locales (en, zh, ja, zh-TW).
